backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
joblib==1.1.0
loguru==0.6.0
nltk==3.7
numpy==1.22.3
openpyxl==3.0.10
pandas==1.4.2
psycopg2==2.9.3
pycodestyle==2.8.0
pymystem3==0.2.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
win32-setctime==1.1.0
